\name{CleanSubsetTS-package}
\alias{CleanSubsetTS-package}
\alias{CleanSubsetTS}
\docType{package}
\title{
\packageTitle{CleanSubsetTS}
}
\description{
\packageDescription{CleanSubsetTS}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{CleanSubsetTS}

Maintainer: \packageMaintainer{CleanSubsetTS}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = CleanSubsetTS)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{CleanSubsetTS}}
%\section{Documentation Index}{\packageIndices{CleanSubsetTS}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
